Create navigator, split into 2 files
